move default implementation to trait
