+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/serializer/ModuleSerializer.scala
@@ -54,79 +54,6 @@ object ModuleSerializer extends ModuleSerializable{
   init
 
   /**
-   * Default deserialization using reflection
-   * @param model serialized protobuf module instace
-   * @return BigDL module instance with linkages with other modules
-   */
-  override def loadModule[T: ClassTag](model : BigDLModule)
-    (implicit ev: TensorNumeric[T]) : ModuleData[T] = {
-
-    val evidence = scala.reflect.classTag[T]
-    val modelAttributes = model.getAttrMap
-    val moduleType = model.getModuleType
-    val cls = ModuleSerializer.getModuleClsByType(moduleType)
-    val constructorMirror = getCostructorMirror(cls)
-    val constructorFullParams = constructorMirror.symbol.paramss
-    val args = new Array[Object](constructorFullParams(0).size + constructorFullParams(1).size)
-    var i = 0;
-    constructorFullParams.foreach(map => {
-      map.foreach(param => {
-        val name = param.name.decodedName.toString
-        val ptype = param.typeSignature
-        if (ptype.toString == "scala.reflect.ClassTag[T]") {
-          args(i) = evidence
-        } else if (ptype.toString ==
-          tensorNumericType.toString) {
-          args(i) = ev
-        } else {
-          require(modelAttributes.containsKey(name), s"$name value cannot be found")
-          val attribute = modelAttributes.get(name)
-          val value = DataConverter.getAttributeValue(attribute)
-          args(i) = value
-        }
-        i+= 1
-      })
-    })
-    val module = constructorMirror.apply(args : _*).
-      asInstanceOf[AbstractModule[Activity, Activity, T]]
-    createBigDLModule(model, module)
-  }
-
-  /**
-   *  Default serialization using reflection
-   *  @param module BigDL module instance with linkages with other modules
-   *  @return serialized protobuf module instace
-   */
-  override def serializeModule[T: ClassTag](module : ModuleData[T])
-                                           (implicit ev: TensorNumeric[T]) : BigDLModule = {
-    val bigDLModelBuilder = BigDLModule.newBuilder
-    val cls = module.module.getClass
-    val moduleType = getModuleTypeByCls(cls)
-    bigDLModelBuilder.setModuleType(moduleType)
-    val fullParams = getCostructorMirror(cls).symbol.paramss
-    val clsTag = scala.reflect.classTag[T]
-    val constructorParams = fullParams(0)
-    constructorParams.foreach(param => {
-      val paramName = param.name.decodedName.toString
-      var ptype = param.typeSignature
-      val attrBuilder = AttrValue.newBuilder
-      // For some modules, fields are declared inside but passed to Super directly
-      var field : Field = null
-      try {
-        field = cls.getDeclaredField(paramName)
-      } catch {
-        case e : NoSuchFieldException =>
-          field = cls.getSuperclass.getDeclaredField(paramName)
-      }
-      field.setAccessible(true)
-      val fieldValue = field.get(module.module)
-      DataConverter.setAttributeValue(attrBuilder, fieldValue, ptype)
-      bigDLModelBuilder.putAttr(paramName, attrBuilder.build)
-    })
-    createSerializeBigDLModule(bigDLModelBuilder, module)
-  }
-
-  /**
    * Serialization entry for all modules based on corresponding class instance of module
    * @param bigDLModule : BigDL module to be serialized
    * @return protobuf format module instance
@@ -164,17 +91,18 @@ object ModuleSerializer extends ModuleSerializable{
     deserializerMaps(moduleType) = serializer
   }
 
-  private def getModuleClsByType(moduleType : String) : Class[_] = {
+  private[serializer] def getModuleClsByType(moduleType : String) : Class[_] = {
     require(moduleMaps.contains(moduleType), s"$moduleType is not supported")
     moduleMaps(moduleType)
   }
 
-  private def getModuleTypeByCls(cls : Class[_]) : String = {
+  private[serializer] def getModuleTypeByCls(cls : Class[_]) : String = {
     require(classMaps.contains(cls), s"$cls is not supported")
     classMaps(cls)
   }
 
-  private def getCostructorMirror[T : ClassTag](cls : Class[_]) : universe.MethodMirror = {
+  private[serializer] def getCostructorMirror[T : ClassTag](cls : Class[_]):
+    universe.MethodMirror = {
 
     val clsSymbol = runtimeMirror.classSymbol(cls)
     val cm = runtimeMirror.reflectClass(clsSymbol)
